+++ NEWS.md
@@ -2,6 +2,8 @@
 
 CHANGES
 
+* `makepredictcall(<dw_transformer>)` now supports `winsorize()`, `reverse()`, and `slide()`.
+
 * `data_read()` now also reads zip-files from URLs (#682).
 
 * `data_tabulate()` now returns an attribute "by" with the